requirements.txt
py-cord==2.7.0
python-dotenv